Solution assignment 3
I implemented the assignment using a graph traversal with branch pruning. Each letter is a node and the edges correspond to diagonal & straight neighbours of each letter in the input matrix. Finding words corresponds to a graph traversal. For efficiency, I prune the traversal branches when the word created so far does not correspond to a prefix in the dictionary. For the traversal, I chose depth-first search to arguably find full words faster than BFS. The worst-case complexity is O(rows * cols) , while the average-case complexity depends on the number of prefixes in the dictionary. In fact, how exactly should I reason about the average-case complexity?